Fix bug in find_best_partitionings()
Original code doesn't give any chance to the last partitioning within partition search limit (only for dual planes case). The confusing part in code is that partition_search_limit is really a number of partitions to test within a range [1..1024] and defacto_search_limit is an index of last partition to be tested and potentially could be [0..1023] (still in actual code it could never be 0, so only [1..1023] is possible).

This change should improve quality for rare cases where last partitioning within search limit is happened to be the best.