Optimize trials of using 2/3 partitions and two planes of weights (#147)
In the original code running trials for 2 or 3 partitions using two planes of weights, two different color components were tested as candidates for the second weight plane. The second component is rarely a good candidate, so the IQ benefits of this are tiny given the processing cost.

For the Kodak test set with -medium compression we get an average of 8% performance improvement, for negligible quality loss (< 0.01 dB).

Coding time:    Mean: +1.08x    Std: 0.03x
Image quality:  Mean: -0.00 dB  Std: 0.00 dB

There is a larger quality impact for -through and -exhaustive - although it is still very small (0.03 dB) - in particular for normal maps. However, in reality no developers are really using those search modes in production as they are far too slow so this seems acceptable.