Add diagnostic tracing tool (#204)
This PR adds support for tracing the compressor execution.

This is implemented using a special build-variant, which can include additional processing and forces single threaded execution. Each compression will generate a JSON file containing the trace file. The trace file is relatively raw, due to the need to keep the tracing stateless. A python tool is provided to post-process the trace and allow further analysis of the data.