Conditionally skip gathers in comp_ideal_weights_for_decimation
The weight_error_scale array is constant around half of the time,
so we often run expensive gathers in AVX2 builds to fetch identical
data.

This change makes the gather in the AVX2 path conditional on
whether the array is constant or not, skipping the gather when
its not needed. This only helps AVX2 builds, and improves overall
runtome by 1-2% for -medium.