Use vtbl/pshufb for small tables (#338)
The current code uses gathers for data-dependent data lookups. These perform "okay" on AVX2, but fall back to a scalar emulation of a gather on SSE and NEON. For small tables we can fit into a couple registers, it is preferable to replace gathers with an in-register table lookup (even on AVX2). NEON has this natively with vtbl, but I'd been struggling to find a good abstraction for this on x86. Ryg and Castano highlighted a nice way to emulate this using pshufb by xor'ing the table entries (https://twitter.com/castano/status/1532381774541512705) which nicely makes this a drop-in alternative to NEON vtbl if we can inject a little bit of table preprocessing for x86 when using pshufb.