Add more control over high quality partition searches (#379)
This PR adds multiple improvements to how we handle partitioning searches during encoding.

Firstly, it adds the ability to compute error estimates for different numbers of partitioning for each of 2/3/4 partitions. This heuristic existed before, but used the same count for all three. This allows us to bias the search towards the 2 partition case, where it is more likely to be useful, and away from the 4 partition case.

Secondly, it adds the ability for 2/3/4 partition searches to run full coding trials on a variable number of partition candidates; again, with a unique setting for each partition count. Prior to this they could only ever test exactly two based on error estimates. This improves IQ, as error estimates cannot accurately factor in quantization, at the expense of significantly more coding time.

Finally, a new -verythorough preset is introduced to provide a drop-in for the old -exhaustive (as the new -exhaustive is 2-3x slower). It benefits from the new heuristics to improve both IQ and performance over the previous -exhaustive setting.

For -verythorough 4x4 blocks the Kodak suite is not 0.08 dB higher IQ, and run at 1.08x speed. 6x6 blocks have 0.08 dB higher IQ, and run at 1.02x speed.

For -exhaustive 4x4 blocks the Kodak suite is now 0.13 dB higher IQ and runs at 0.4x speed. 6x6 blocks have 0.10 dB higher IQ and run at 0.3x speed.