Remove folded decimation_info weight arrays
Falling back to the non-folded arrays is marginally slower on perfect
memory because of additional indirect loads. However, removing the
folded arrays significantly improves caching which offsets the loss.

This change reduces the context creation time and memory footprint of
the compressor. This is most significant for larger block sizes which
have the most decimation_info structures to create and access.