macOS universal build improvements (#423)
This change provides a new implementation of macOS universal binary builds, which is considerably simpler than the previous integration and more useful in as far as it supports the x86_64h slice for AVX2.

Using the ASTCENC_UNIVERSAL_BUILD will now always build the SSE41, AVX2 and NEON single ISA build variants, and use lipo to package them together into a single binary. When in universal mode only the combined binary is installed; the single ISA binaries are built but are treated as internal intermediate build artefacts.

This is the only ISA support provided by the universal build - you can't build with a subset, and you can't use the NONE or SSE2 variants.