Multi-pack coding and pack vectorization (#424)
This change implements a change so that RGB and RGBA LDR color packing will try multiple packing methods (delta, blue-contract, delta+blue-contract, or normal) and selects the packing with the lowest cumulative endpoint error.

This change also changes vector float_to_int rounding so that it matches the scalar flt2int, which is round-down on tie, not round-to-nearest-even on tie. This causes some minor noise-level IQ differences, but is useful as it means the scalar and vector paths now agree which should makes future vectorization changes bit-exact.

This changes is ~performance neutral with the added vectorization of the LDR RGB and RGBA color packing, for an average of 0.05 dB quality improvement (some images over 0.1 dB)

Fixes #422.